OpenAPIPetstore.API.User

Description

Operations

User

createUser

createUser

Arguments

:: (Consumes CreateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> User
"body" - Created user object

-> OpenAPIPetstoreRequest CreateUser contentType NoContent MimeNoContent

POST /user

Create user

This can only be done by the logged in user.

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: (Consumes CreateUsersWithArrayInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Body
"body" - List of user object

-> OpenAPIPetstoreRequest CreateUsersWithArrayInput contentType NoContent MimeNoContent

POST /user/createWithArray

Creates list of users with given input array

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: (Consumes CreateUsersWithListInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Body
"body" - List of user object

-> OpenAPIPetstoreRequest CreateUsersWithListInput contentType NoContent MimeNoContent

POST /user/createWithList

Creates list of users with given input array

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Username
"username" - The name that needs to be deleted

-> OpenAPIPetstoreRequest DeleteUser MimeNoContent NoContent MimeNoContent

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

data DeleteUser

getUserByName

getUserByName

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The name that needs to be fetched. Use user1 for testing.

-> OpenAPIPetstoreRequest GetUserByName MimeNoContent User accept

GET /user/{username}

Get user by user name

data GetUserByName

loginUser

loginUser

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The user name for login

-> Password
"password" - The password for login in clear text

-> OpenAPIPetstoreRequest LoginUser MimeNoContent Text accept

GET /user/login

Logs user into the system

data LoginUser

logoutUser

logoutUser :: OpenAPIPetstoreRequest LogoutUser MimeNoContent NoContent MimeNoContent

GET /user/logout

Logs out current logged in user session

data LogoutUser

updateUser

updateUser

Arguments

:: (Consumes UpdateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> User
"body" - Updated user object

-> Username
"username" - name that need to be deleted

-> OpenAPIPetstoreRequest UpdateUser contentType NoContent MimeNoContent

PUT /user/{username}

Updated user

This can only be done by the logged in user.

data UpdateUser