OpenAPIPetstore.Core

Description

OpenAPIPetstoreConfig

data OpenAPIPetstoreConfig

Constructors

OpenAPIPetstoreConfig

Fields

configHost :: ByteString

host supplied in the Request

configUserAgent :: Text

user-agent supplied in the Request

configLogExecWithContext :: LogExecWithContext

Run a block using a Logger instance

configLogContext :: LogContext

Configures the logger

configAuthMethods :: [AnyAuthMethod]

List of configured auth methods

configValidateAuthMethods :: Bool

throw exceptions if auth methods are not configured

configQueryExtraUnreserved :: ByteString

Configures additional querystring characters which must not be URI encoded, e.g. + or :

newConfig :: IO OpenAPIPetstoreConfig

constructs a default OpenAPIPetstoreConfig

configHost:

http://petstore.swagger.io:80/v2

configUserAgent:

"openapi-petstore/0.1.0.0"

addAuthMethod :: AuthMethod auth => OpenAPIPetstoreConfig -> auth -> OpenAPIPetstoreConfig

updates config use AuthMethod on matching requests

withStdoutLogging :: OpenAPIPetstoreConfig -> IO OpenAPIPetstoreConfig

updates the config to use stdout logging

withStderrLogging :: OpenAPIPetstoreConfig -> IO OpenAPIPetstoreConfig

updates the config to use stderr logging

withNoLogging :: OpenAPIPetstoreConfig -> OpenAPIPetstoreConfig

updates the config to disable logging

OpenAPIPetstoreRequest

data OpenAPIPetstoreRequest req contentType res accept

Represents a request.

Type Variables:

req - request operation

contentType - MimeType associated with request body

res - response model

accept - MimeType associated with response body

Constructors

OpenAPIPetstoreRequest

Fields

rMethod :: Method

Method of OpenAPIPetstoreRequest

rUrlPath :: [ByteString]

Endpoint of OpenAPIPetstoreRequest

rParams :: Params

params of OpenAPIPetstoreRequest

rAuthTypes :: [TypeRep]

types of auth methods

rMethodL :: Lens_' (OpenAPIPetstoreRequest req contentType res accept) Method

rMethod Lens

rUrlPathL :: Lens_' (OpenAPIPetstoreRequest req contentType res accept) [ByteString]

rUrlPath Lens

rParamsL :: Lens_' (OpenAPIPetstoreRequest req contentType res accept) Params

rParams Lens

rAuthTypesL :: Lens_' (OpenAPIPetstoreRequest req contentType res accept) [TypeRep]

rParams Lens

HasBodyParam

class HasBodyParam req param where

Designates the body parameter of a request

Minimal complete definition

Nothing

Methods

setBodyParam :: forall contentType res accept. (Consumes req contentType, MimeRender contentType param) => OpenAPIPetstoreRequest req contentType res accept -> param -> OpenAPIPetstoreRequest req contentType res accept

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: OpenAPIPetstoreRequest req contentType res accept -> param -> OpenAPIPetstoreRequest req contentType res accept

Apply an optional parameter to a request

(-&-) :: OpenAPIPetstoreRequest req contentType res accept -> param -> OpenAPIPetstoreRequest req contentType res accept infixl 2

infix operator / alias for addOptionalParam

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

paramsQueryL :: Lens_' Params Query

paramsQuery Lens

paramsHeadersL :: Lens_' Params RequestHeaders

paramsHeaders Lens

paramsBodyL :: Lens_' Params ParamBody

paramsBody Lens

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyB ByteString

ParamBodyBL ByteString

ParamBodyFormUrlEncoded Form

ParamBodyMultipartFormData [Part]

OpenAPIPetstoreRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> OpenAPIPetstoreRequest req contentType res accept
req: Request Type, res: Response Type

_mkParams :: Params

setHeader :: OpenAPIPetstoreRequest req contentType res accept -> [Header] -> OpenAPIPetstoreRequest req contentType res accept

addHeader :: OpenAPIPetstoreRequest req contentType res accept -> [Header] -> OpenAPIPetstoreRequest req contentType res accept

removeHeader :: OpenAPIPetstoreRequest req contentType res accept -> [HeaderName] -> OpenAPIPetstoreRequest req contentType res accept

_setContentTypeHeader :: forall req contentType res accept. MimeType contentType => OpenAPIPetstoreRequest req contentType res accept -> OpenAPIPetstoreRequest req contentType res accept

_setAcceptHeader :: forall req contentType res accept. MimeType accept => OpenAPIPetstoreRequest req contentType res accept -> OpenAPIPetstoreRequest req contentType res accept

setQuery :: OpenAPIPetstoreRequest req contentType res accept -> [QueryItem] -> OpenAPIPetstoreRequest req contentType res accept

addQuery :: OpenAPIPetstoreRequest req contentType res accept -> [QueryItem] -> OpenAPIPetstoreRequest req contentType res accept

addForm :: OpenAPIPetstoreRequest req contentType res accept -> Form -> OpenAPIPetstoreRequest req contentType res accept

_addMultiFormPart :: OpenAPIPetstoreRequest req contentType res accept -> Part -> OpenAPIPetstoreRequest req contentType res accept

_setBodyBS :: OpenAPIPetstoreRequest req contentType res accept -> ByteString -> OpenAPIPetstoreRequest req contentType res accept

_setBodyLBS :: OpenAPIPetstoreRequest req contentType res accept -> ByteString -> OpenAPIPetstoreRequest req contentType res accept

_hasAuthType :: AuthMethod authMethod => OpenAPIPetstoreRequest req contentType res accept -> Proxy authMethod -> OpenAPIPetstoreRequest req contentType res accept

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

toPartialEscapeQuery :: ByteString -> Query -> PartialEscapeQuery

OpenAPI CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

AuthMethods

class Typeable a => AuthMethod a where

Provides a method to apply auth methods to requests

Methods

applyAuthMethod :: OpenAPIPetstoreConfig -> a -> OpenAPIPetstoreRequest req contentType res accept -> IO (OpenAPIPetstoreRequest req contentType res accept)

data AnyAuthMethod

An existential wrapper for any AuthMethod

Constructors

forall a.AuthMethod a => AnyAuthMethod a

data AuthMethodException

indicates exceptions related to AuthMethods

Constructors

AuthMethodException String

_applyAuthMethods :: OpenAPIPetstoreRequest req contentType res accept -> OpenAPIPetstoreConfig -> IO (OpenAPIPetstoreRequest req contentType res accept)

apply all matching AuthMethods in config to request

Utils

_omitNulls :: [(Text, Value)] -> Value

Removes Null fields. (OpenAPI-Specification 2.0 does not allow Null in JSON)

_toFormItem :: (ToHttpApiData a, Functor f) => t -> f a -> f (t, [Text])

Encodes fields using WH.toQueryParam

_emptyToNothing :: Maybe String -> Maybe String

Collapse (Just "") to Nothing

_memptyToNothing :: (Monoid a, Eq a) => Maybe a -> Maybe a

Collapse (Just mempty) to Nothing

DateTime Formatting

newtype DateTime

Constructors

DateTime

Fields

unDateTime :: UTCTime

_readDateTime :: (MonadFail m, Alternative m) => String -> m DateTime

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, MonadFail m, Alternative m) => String -> m t

parse an ISO8601 date-time string

Date Formatting

newtype Date

Constructors

Date

Fields

unDate :: Day

_readDate :: MonadFail m => String -> m Date

TI.parseTimeM True TI.defaultTimeLocale "%Y-%m-%d"

_showDate :: FormatTime t => t -> String

TI.formatTime TI.defaultTimeLocale "%Y-%m-%d"

Byte/Binary Formatting

newtype ByteArray

base64 encoded characters

Constructors

ByteArray

Fields

unByteArray :: ByteString

_readByteArray :: MonadFail m => Text -> m ByteArray

read base64 encoded characters

_showByteArray :: ByteArray -> Text

show base64 encoded characters

newtype Binary

any sequence of octets

Constructors

Binary

Fields

unBinary :: ByteString

_readBinaryBase64 :: MonadFail m => Text -> m Binary

_showBinaryBase64 :: Binary -> Text

Lens Type Aliases

type Lens_' s a = Lens_ s s a a

type Lens_ s t a b = forall (f :: Type -> Type). Functor f => (a -> f b) -> s -> f t