OpenAPIPetstore.Logging

Description

Logging functions

Type Aliases (for compatibility)

type LogExecWithContext = forall m a. MonadIO m => LogContext -> LogExec m a

Runs a Katip logging block with the Log environment

type LogExec m a = KatipT m a -> m a

A Katip logging block

type LogContext = LogEnv

A Katip Log environment

type LogLevel = Severity

A Katip Log severity

default logger

initLogContext :: IO LogContext

the default log environment

runDefaultLogExecWithContext :: LogExecWithContext

Runs a Katip logging block with the Log environment

stdout logger

stdoutLoggingExec :: LogExecWithContext

Runs a Katip logging block with the Log environment

stdoutLoggingContext :: LogContext -> IO LogContext

A Katip Log environment which targets stdout

stderr logger

stderrLoggingExec :: LogExecWithContext

Runs a Katip logging block with the Log environment

stderrLoggingContext :: LogContext -> IO LogContext

A Katip Log environment which targets stderr

Null logger

runNullLogExec :: LogExecWithContext

Disables Katip logging

Log Msg

_log :: (Applicative m, Katip m) => Text -> LogLevel -> Text -> m ()

Log a katip message

Log Exceptions

logExceptions :: (Katip m, MonadCatch m, Applicative m) => Text -> m a -> m a

re-throws exceptions after logging them

Log Level

levelInfo :: LogLevel

levelError :: LogLevel

levelDebug :: LogLevel