OpenAPIPetstore.Model

Description

Parameter newtypes

AdditionalMetadata

newtype AdditionalMetadata

Constructors

AdditionalMetadata

Fields

unAdditionalMetadata :: Text

ApiKey

newtype ApiKey

Constructors

ApiKey

Fields

unApiKey :: Text

Body

newtype Body

Constructors

Body

Fields

unBody :: [User]

BodyBool

newtype BodyBool

Constructors

BodyBool

Fields

unBodyBool :: Bool

BodyDouble

newtype BodyDouble

Constructors

BodyDouble

Fields

unBodyDouble :: Double

BodyText

newtype BodyText

Constructors

BodyText

Fields

unBodyText :: Text

BooleanGroup

newtype BooleanGroup

Constructors

BooleanGroup

Fields

unBooleanGroup :: Bool

Byte

newtype Byte

Constructors

Byte

Fields

unByte :: ByteArray

Callback

newtype Callback

Constructors

Callback

Fields

unCallback :: Text

Context

newtype Context

Constructors

Context

Fields

unContext :: [Text]

EnumFormString

newtype EnumFormString

Constructors

EnumFormString

Fields

unEnumFormString :: E'EnumFormString

EnumFormStringArray

newtype EnumFormStringArray

Constructors

EnumFormStringArray

Fields

unEnumFormStringArray :: [E'EnumFormStringArray]

EnumHeaderString

newtype EnumHeaderString

Constructors

EnumHeaderString

Fields

unEnumHeaderString :: E'EnumFormString

EnumHeaderStringArray

newtype EnumHeaderStringArray

Constructors

EnumHeaderStringArray

Fields

unEnumHeaderStringArray :: [E'EnumFormStringArray]

EnumQueryDouble

newtype EnumQueryDouble

Constructors

EnumQueryDouble

Fields

unEnumQueryDouble :: E'EnumNumber

EnumQueryInteger

newtype EnumQueryInteger

Constructors

EnumQueryInteger

Fields

unEnumQueryInteger :: E'EnumQueryInteger

EnumQueryString

newtype EnumQueryString

Constructors

EnumQueryString

Fields

unEnumQueryString :: E'EnumFormString

EnumQueryStringArray

newtype EnumQueryStringArray

Constructors

EnumQueryStringArray

Fields

unEnumQueryStringArray :: [E'EnumFormStringArray]

File2

newtype File2

Constructors

File2

Fields

unFile2 :: FilePath

Http

newtype Http

Constructors

Http

Fields

unHttp :: [Text]

Int32

newtype Int32

Constructors

Int32

Fields

unInt32 :: Int

Int64

newtype Int64

Constructors

Int64

Fields

unInt64 :: Integer

Int64Group

newtype Int64Group

Constructors

Int64Group

Fields

unInt64Group :: Integer

Ioutil

newtype Ioutil

Constructors

Ioutil

Fields

unIoutil :: [Text]

Name2

newtype Name2

Constructors

Name2

Fields

unName2 :: Text

Number

newtype Number

Constructors

Number

Fields

unNumber :: Double

OrderId

newtype OrderId

Constructors

OrderId

Fields

unOrderId :: Integer

OrderIdText

newtype OrderIdText

Constructors

OrderIdText

Fields

unOrderIdText :: Text

Param

newtype Param

Constructors

Param

Fields

unParam :: Text

Param2

newtype Param2

Constructors

Param2

Fields

unParam2 :: Text

ParamBinary

newtype ParamBinary

Constructors

ParamBinary

Fields

unParamBinary :: FilePath

ParamDate

newtype ParamDate

Constructors

ParamDate

Fields

unParamDate :: Date

ParamDateTime

newtype ParamDateTime

Constructors

ParamDateTime

Fields

unParamDateTime :: DateTime

ParamDouble

newtype ParamDouble

Constructors

ParamDouble

Fields

unParamDouble :: Double

ParamFloat

newtype ParamFloat

Constructors

ParamFloat

Fields

unParamFloat :: Float

ParamInteger

newtype ParamInteger

Constructors

ParamInteger

Fields

unParamInteger :: Int

ParamMapMapStringText

newtype ParamMapMapStringText

Constructors

ParamMapMapStringText

Fields

unParamMapMapStringText :: Map String Text

ParamString

newtype ParamString

Constructors

ParamString

Fields

unParamString :: Text

Password

newtype Password

Constructors

Password

Fields

unPassword :: Text

PatternWithoutDelimiter

newtype PatternWithoutDelimiter

Constructors

PatternWithoutDelimiter

Fields

unPatternWithoutDelimiter :: Text

PetId

newtype PetId

Constructors

PetId

Fields

unPetId :: Integer

Pipe

newtype Pipe

Constructors

Pipe

Fields

unPipe :: [Text]

Query

newtype Query

Constructors

Query

Fields

unQuery :: Text

RequiredBooleanGroup

newtype RequiredBooleanGroup

Constructors

RequiredBooleanGroup

Fields

unRequiredBooleanGroup :: Bool

RequiredFile

newtype RequiredFile

Constructors

RequiredFile

Fields

unRequiredFile :: FilePath

RequiredInt64Group

newtype RequiredInt64Group

Constructors

RequiredInt64Group

Fields

unRequiredInt64Group :: Integer

RequiredStringGroup

newtype RequiredStringGroup

Constructors

RequiredStringGroup

Fields

unRequiredStringGroup :: Int

Status

newtype Status

Constructors

Status

Fields

unStatus :: [E'Status2]

StatusText

newtype StatusText

Constructors

StatusText

Fields

unStatusText :: Text

StringGroup

newtype StringGroup

Constructors

StringGroup

Fields

unStringGroup :: Int

Tags

newtype Tags

Constructors

Tags

Fields

unTags :: [Text]

Url

newtype Url

Constructors

Url

Fields

unUrl :: [Text]

Username

newtype Username

Constructors

Username

Fields

unUsername :: Text

Models

AdditionalPropertiesAnyType

data AdditionalPropertiesAnyType

AdditionalPropertiesAnyType

Constructors

AdditionalPropertiesAnyType

Fields

additionalPropertiesAnyTypeName :: !(Maybe Text)

"name"

mkAdditionalPropertiesAnyType :: AdditionalPropertiesAnyType

Construct a value of type AdditionalPropertiesAnyType (by applying it's required fields, if any)

AdditionalPropertiesArray

data AdditionalPropertiesArray

AdditionalPropertiesArray

Constructors

AdditionalPropertiesArray

Fields

additionalPropertiesArrayName :: !(Maybe Text)

"name"

mkAdditionalPropertiesArray :: AdditionalPropertiesArray

Construct a value of type AdditionalPropertiesArray (by applying it's required fields, if any)

AdditionalPropertiesBoolean

data AdditionalPropertiesBoolean

AdditionalPropertiesBoolean

Constructors

AdditionalPropertiesBoolean

Fields

additionalPropertiesBooleanName :: !(Maybe Text)

"name"

mkAdditionalPropertiesBoolean :: AdditionalPropertiesBoolean

Construct a value of type AdditionalPropertiesBoolean (by applying it's required fields, if any)

AdditionalPropertiesClass

data AdditionalPropertiesClass

AdditionalPropertiesClass

Constructors

AdditionalPropertiesClass

Fields

additionalPropertiesClassMapString :: !(Maybe (Map String Text))

"map_string"

additionalPropertiesClassMapNumber :: !(Maybe (Map String Double))

"map_number"

additionalPropertiesClassMapInteger :: !(Maybe (Map String Int))

"map_integer"

additionalPropertiesClassMapBoolean :: !(Maybe (Map String Bool))

"map_boolean"

additionalPropertiesClassMapArrayInteger :: !(Maybe (Map String [Int]))

"map_array_integer"

additionalPropertiesClassMapArrayAnytype :: !(Maybe (Map String [Value]))

"map_array_anytype"

additionalPropertiesClassMapMapString :: !(Maybe (Map String (Map String Text)))

"map_map_string"

additionalPropertiesClassMapMapAnytype :: !(Maybe (Map String (Map String Value)))

"map_map_anytype"

additionalPropertiesClassAnytype1 :: !(Maybe Value)

"anytype_1"

additionalPropertiesClassAnytype2 :: !(Maybe Value)

"anytype_2"

additionalPropertiesClassAnytype3 :: !(Maybe Value)

"anytype_3"

mkAdditionalPropertiesClass :: AdditionalPropertiesClass

Construct a value of type AdditionalPropertiesClass (by applying it's required fields, if any)

AdditionalPropertiesInteger

data AdditionalPropertiesInteger

AdditionalPropertiesInteger

Constructors

AdditionalPropertiesInteger

Fields

additionalPropertiesIntegerName :: !(Maybe Text)

"name"

mkAdditionalPropertiesInteger :: AdditionalPropertiesInteger

Construct a value of type AdditionalPropertiesInteger (by applying it's required fields, if any)

AdditionalPropertiesNumber

data AdditionalPropertiesNumber

AdditionalPropertiesNumber

Constructors

AdditionalPropertiesNumber

Fields

additionalPropertiesNumberName :: !(Maybe Text)

"name"

mkAdditionalPropertiesNumber :: AdditionalPropertiesNumber

Construct a value of type AdditionalPropertiesNumber (by applying it's required fields, if any)

AdditionalPropertiesObject

data AdditionalPropertiesObject

AdditionalPropertiesObject

Constructors

AdditionalPropertiesObject

Fields

additionalPropertiesObjectName :: !(Maybe Text)

"name"

mkAdditionalPropertiesObject :: AdditionalPropertiesObject

Construct a value of type AdditionalPropertiesObject (by applying it's required fields, if any)

AdditionalPropertiesString

data AdditionalPropertiesString

AdditionalPropertiesString

Constructors

AdditionalPropertiesString

Fields

additionalPropertiesStringName :: !(Maybe Text)

"name"

mkAdditionalPropertiesString :: AdditionalPropertiesString

Construct a value of type AdditionalPropertiesString (by applying it's required fields, if any)

Animal

data Animal

Animal

Constructors

Animal

Fields

animalClassName :: !Text

Required "className"

animalColor :: !(Maybe Text)

"color"

mkAnimal

Arguments

:: Text
animalClassName

-> Animal

Construct a value of type Animal (by applying it's required fields, if any)

ApiResponse

data ApiResponse

ApiResponse

Constructors

ApiResponse

Fields

apiResponseCode :: !(Maybe Int)

"code"

apiResponseType :: !(Maybe Text)

"type"

apiResponseMessage :: !(Maybe Text)

"message"

mkApiResponse :: ApiResponse

Construct a value of type ApiResponse (by applying it's required fields, if any)

ArrayOfArrayOfNumberOnly

data ArrayOfArrayOfNumberOnly

ArrayOfArrayOfNumberOnly

Constructors

ArrayOfArrayOfNumberOnly

Fields

arrayOfArrayOfNumberOnlyArrayArrayNumber :: !(Maybe [[Double]])

ArrayArrayNumber

mkArrayOfArrayOfNumberOnly :: ArrayOfArrayOfNumberOnly

Construct a value of type ArrayOfArrayOfNumberOnly (by applying it's required fields, if any)

ArrayOfNumberOnly

data ArrayOfNumberOnly

ArrayOfNumberOnly

Constructors

ArrayOfNumberOnly

Fields

arrayOfNumberOnlyArrayNumber :: !(Maybe [Double])

ArrayNumber

mkArrayOfNumberOnly :: ArrayOfNumberOnly

Construct a value of type ArrayOfNumberOnly (by applying it's required fields, if any)

ArrayTest

data ArrayTest

ArrayTest

Constructors

ArrayTest

Fields

arrayTestArrayOfString :: !(Maybe [Text])

"array_of_string"

arrayTestArrayArrayOfInteger :: !(Maybe [[Integer]])

"array_array_of_integer"

arrayTestArrayArrayOfModel :: !(Maybe [[ReadOnlyFirst]])

"array_array_of_model"

mkArrayTest :: ArrayTest

Construct a value of type ArrayTest (by applying it's required fields, if any)

BigCat

data BigCat

BigCat

Constructors

BigCat

Fields

bigCatClassName :: !Text

Required "className"

bigCatColor :: !(Maybe Text)

"color"

bigCatDeclawed :: !(Maybe Bool)

"declawed"

bigCatKind :: !(Maybe E'Kind)

"kind"

mkBigCat

Arguments

:: Text
bigCatClassName

-> BigCat

Construct a value of type BigCat (by applying it's required fields, if any)

BigCatAllOf

data BigCatAllOf

BigCatAllOf

Constructors

BigCatAllOf

Fields

bigCatAllOfKind :: !(Maybe E'Kind)

"kind"

mkBigCatAllOf :: BigCatAllOf

Construct a value of type BigCatAllOf (by applying it's required fields, if any)

Capitalization

data Capitalization

Capitalization

Constructors

Capitalization

Fields

capitalizationSmallCamel :: !(Maybe Text)

"smallCamel"

capitalizationCapitalCamel :: !(Maybe Text)

CapitalCamel

capitalizationSmallSnake :: !(Maybe Text)

"small_Snake"

capitalizationCapitalSnake :: !(Maybe Text)

Capital_Snake

capitalizationScaEthFlowPoints :: !(Maybe Text)

SCA_ETH_Flow_Points

capitalizationAttName :: !(Maybe Text)

ATT_NAME - Name of the pet

mkCapitalization :: Capitalization

Construct a value of type Capitalization (by applying it's required fields, if any)

Cat

data Cat

Cat

Constructors

Cat

Fields

catClassName :: !Text

Required "className"

catColor :: !(Maybe Text)

"color"

catDeclawed :: !(Maybe Bool)

"declawed"

mkCat

Arguments

:: Text
catClassName

-> Cat

Construct a value of type Cat (by applying it's required fields, if any)

CatAllOf

data CatAllOf

CatAllOf

Constructors

CatAllOf

Fields

catAllOfDeclawed :: !(Maybe Bool)

"declawed"

mkCatAllOf :: CatAllOf

Construct a value of type CatAllOf (by applying it's required fields, if any)

Category

data Category

Category

Constructors

Category

Fields

categoryId :: !(Maybe Integer)

"id"

categoryName :: !Text

Required "name"

mkCategory

Arguments

:: Text
categoryName

-> Category

Construct a value of type Category (by applying it's required fields, if any)

ClassModel

data ClassModel

ClassModel Model for testing model with "_class" property

Constructors

ClassModel

Fields

classModelClass :: !(Maybe Text)

"_class"

mkClassModel :: ClassModel

Construct a value of type ClassModel (by applying it's required fields, if any)

Client

data Client

Client

Constructors

Client

Fields

clientClient :: !(Maybe Text)

"client"

mkClient :: Client

Construct a value of type Client (by applying it's required fields, if any)

Dog

data Dog

Dog

Constructors

Dog

Fields

dogClassName :: !Text

Required "className"

dogColor :: !(Maybe Text)

"color"

dogBreed :: !(Maybe Text)

"breed"

mkDog

Arguments

:: Text
dogClassName

-> Dog

Construct a value of type Dog (by applying it's required fields, if any)

DogAllOf

data DogAllOf

DogAllOf

Constructors

DogAllOf

Fields

dogAllOfBreed :: !(Maybe Text)

"breed"

mkDogAllOf :: DogAllOf

Construct a value of type DogAllOf (by applying it's required fields, if any)

EnumArrays

data EnumArrays

EnumArrays

Constructors

EnumArrays

Fields

enumArraysJustSymbol :: !(Maybe E'JustSymbol)

"just_symbol"

enumArraysArrayEnum :: !(Maybe [E'ArrayEnum])

"array_enum"

mkEnumArrays :: EnumArrays

Construct a value of type EnumArrays (by applying it's required fields, if any)

EnumTest

data EnumTest

EnumTest

Constructors

EnumTest

Fields

enumTestEnumString :: !(Maybe E'EnumString)

"enum_string"

enumTestEnumStringRequired :: !E'EnumString

Required "enum_string_required"

enumTestEnumInteger :: !(Maybe E'EnumInteger)

"enum_integer"

enumTestEnumNumber :: !(Maybe E'EnumNumber)

"enum_number"

enumTestOuterEnum :: !(Maybe OuterEnum)

"outerEnum"

mkEnumTest

Arguments

:: E'EnumString
enumTestEnumStringRequired

-> EnumTest

Construct a value of type EnumTest (by applying it's required fields, if any)

File

data File

File Must be named File for test.

Constructors

File

Fields

fileSourceUri :: !(Maybe Text)

"sourceURI" - Test capitalization

mkFile :: File

Construct a value of type File (by applying it's required fields, if any)

FileSchemaTestClass

data FileSchemaTestClass

FileSchemaTestClass

Constructors

FileSchemaTestClass

Fields

fileSchemaTestClassFile :: !(Maybe File)

"file"

fileSchemaTestClassFiles :: !(Maybe [File])

"files"

mkFileSchemaTestClass :: FileSchemaTestClass

Construct a value of type FileSchemaTestClass (by applying it's required fields, if any)

FormatTest

data FormatTest

FormatTest

Constructors

FormatTest

Fields

formatTestInteger :: !(Maybe Int)

"integer"

formatTestInt32 :: !(Maybe Int)

"int32"

formatTestInt64 :: !(Maybe Integer)

"int64"

formatTestNumber :: !Double

Required "number"

formatTestFloat :: !(Maybe Float)

"float"

formatTestDouble :: !(Maybe Double)

"double"

formatTestString :: !(Maybe Text)

"string"

formatTestByte :: !ByteArray

Required "byte"

formatTestBinary :: !(Maybe FilePath)

"binary"

formatTestDate :: !Date

Required "date"

formatTestDateTime :: !(Maybe DateTime)

"dateTime"

formatTestUuid :: !(Maybe Text)

"uuid"

formatTestPassword :: !Text

Required "password"

formatTestBigDecimal :: !(Maybe Double)

BigDecimal

mkFormatTest

Arguments

:: Double
formatTestNumber

-> ByteArray
formatTestByte

-> Date
formatTestDate

-> Text
formatTestPassword

-> FormatTest

Construct a value of type FormatTest (by applying it's required fields, if any)

HasOnlyReadOnly

data HasOnlyReadOnly

HasOnlyReadOnly

Constructors

HasOnlyReadOnly

Fields

hasOnlyReadOnlyBar :: !(Maybe Text)

ReadOnly "bar"

hasOnlyReadOnlyFoo :: !(Maybe Text)

ReadOnly "foo"

mkHasOnlyReadOnly :: HasOnlyReadOnly

Construct a value of type HasOnlyReadOnly (by applying it's required fields, if any)

MapTest

data MapTest

MapTest

Constructors

MapTest

Fields

mapTestMapMapOfString :: !(Maybe (Map String (Map String Text)))

"map_map_of_string"

mapTestMapOfEnumString :: !(Maybe (Map String E'Inner))

"map_of_enum_string"

mapTestDirectMap :: !(Maybe (Map String Bool))

"direct_map"

mapTestIndirectMap :: !(Maybe (Map String Bool))

"indirect_map"

mkMapTest :: MapTest

Construct a value of type MapTest (by applying it's required fields, if any)

MixedPropertiesAndAdditionalPropertiesClass

data MixedPropertiesAndAdditionalPropertiesClass

MixedPropertiesAndAdditionalPropertiesClass

Constructors

MixedPropertiesAndAdditionalPropertiesClass

Fields

mixedPropertiesAndAdditionalPropertiesClassUuid :: !(Maybe Text)

"uuid"

mixedPropertiesAndAdditionalPropertiesClassDateTime :: !(Maybe DateTime)

"dateTime"

mixedPropertiesAndAdditionalPropertiesClassMap :: !(Maybe (Map String Animal))

"map"

mkMixedPropertiesAndAdditionalPropertiesClass :: MixedPropertiesAndAdditionalPropertiesClass

Construct a value of type MixedPropertiesAndAdditionalPropertiesClass (by applying it's required fields, if any)

Model200Response

data Model200Response

Model200Response Model for testing model name starting with number

Constructors

Model200Response

Fields

model200ResponseName :: !(Maybe Int)

"name"

model200ResponseClass :: !(Maybe Text)

"class"

mkModel200Response :: Model200Response

Construct a value of type Model200Response (by applying it's required fields, if any)

ModelList

data ModelList

ModelList

Constructors

ModelList

Fields

modelList123list :: !(Maybe Text)

"123-list"

mkModelList :: ModelList

Construct a value of type ModelList (by applying it's required fields, if any)

ModelReturn

data ModelReturn

ModelReturn Model for testing reserved words

Constructors

ModelReturn

Fields

modelReturnReturn :: !(Maybe Int)

"return"

mkModelReturn :: ModelReturn

Construct a value of type ModelReturn (by applying it's required fields, if any)

Name

data Name

Name Model for testing model name same as property name

Constructors

Name

Fields

nameName :: !Int

Required "name"

nameSnakeCase :: !(Maybe Int)

ReadOnly "snake_case"

nameProperty :: !(Maybe Text)

"property"

name123number :: !(Maybe Int)

ReadOnly "123Number"

mkName

Arguments

:: Int
nameName

-> Name

Construct a value of type Name (by applying it's required fields, if any)

NumberOnly

data NumberOnly

NumberOnly

Constructors

NumberOnly

Fields

numberOnlyJustNumber :: !(Maybe Double)

JustNumber

mkNumberOnly :: NumberOnly

Construct a value of type NumberOnly (by applying it's required fields, if any)

Order

data Order

Order

Constructors

Order

Fields

orderId :: !(Maybe Integer)

"id"

orderPetId :: !(Maybe Integer)

"petId"

orderQuantity :: !(Maybe Int)

"quantity"

orderShipDate :: !(Maybe DateTime)

"shipDate"

orderStatus :: !(Maybe E'Status)

"status" - Order Status

orderComplete :: !(Maybe Bool)

"complete"

mkOrder :: Order

Construct a value of type Order (by applying it's required fields, if any)

OuterComposite

data OuterComposite

OuterComposite

Constructors

OuterComposite

Fields

outerCompositeMyNumber :: !(Maybe Double)

"my_number"

outerCompositeMyString :: !(Maybe Text)

"my_string"

outerCompositeMyBoolean :: !(Maybe Bool)

"my_boolean"

mkOuterComposite :: OuterComposite

Construct a value of type OuterComposite (by applying it's required fields, if any)

Pet

data Pet

Pet

Constructors

Pet

Fields

petId :: !(Maybe Integer)

"id"

petCategory :: !(Maybe Category)

"category"

petName :: !Text

Required "name"

petPhotoUrls :: ![Text]

Required "photoUrls"

petTags :: !(Maybe [Tag])

"tags"

petStatus :: !(Maybe E'Status2)

"status" - pet status in the store

mkPet

Arguments

:: Text
petName

-> [Text]
petPhotoUrls

-> Pet

Construct a value of type Pet (by applying it's required fields, if any)

ReadOnlyFirst

data ReadOnlyFirst

ReadOnlyFirst

Constructors

ReadOnlyFirst

Fields

readOnlyFirstBar :: !(Maybe Text)

ReadOnly "bar"

readOnlyFirstBaz :: !(Maybe Text)

"baz"

mkReadOnlyFirst :: ReadOnlyFirst

Construct a value of type ReadOnlyFirst (by applying it's required fields, if any)

SpecialModelName

data SpecialModelName

SpecialModelName

Constructors

SpecialModelName

Fields

specialModelNameSpecialPropertyName :: !(Maybe Integer)

"$special[property.name]"

mkSpecialModelName :: SpecialModelName

Construct a value of type SpecialModelName (by applying it's required fields, if any)

Tag

data Tag

Tag

Constructors

Tag

Fields

tagId :: !(Maybe Integer)

"id"

tagName :: !(Maybe Text)

"name"

mkTag :: Tag

Construct a value of type Tag (by applying it's required fields, if any)

TypeHolderDefault

data TypeHolderDefault

TypeHolderDefault

Constructors

TypeHolderDefault

Fields

typeHolderDefaultStringItem :: !Text

Required "string_item"

typeHolderDefaultNumberItem :: !Double

Required "number_item"

typeHolderDefaultIntegerItem :: !Int

Required "integer_item"

typeHolderDefaultBoolItem :: !Bool

Required "bool_item"

typeHolderDefaultArrayItem :: ![Int]

Required "array_item"

mkTypeHolderDefault

Arguments

:: Text
typeHolderDefaultStringItem

-> Double
typeHolderDefaultNumberItem

-> Int
typeHolderDefaultIntegerItem

-> Bool
typeHolderDefaultBoolItem

-> [Int]
typeHolderDefaultArrayItem

-> TypeHolderDefault

Construct a value of type TypeHolderDefault (by applying it's required fields, if any)

TypeHolderExample

data TypeHolderExample

TypeHolderExample

Constructors

TypeHolderExample

Fields

typeHolderExampleStringItem :: !Text

Required "string_item"

typeHolderExampleNumberItem :: !Double

Required "number_item"

typeHolderExampleFloatItem :: !Float

Required "float_item"

typeHolderExampleIntegerItem :: !Int

Required "integer_item"

typeHolderExampleBoolItem :: !Bool

Required "bool_item"

typeHolderExampleArrayItem :: ![Int]

Required "array_item"

mkTypeHolderExample

Arguments

:: Text
typeHolderExampleStringItem

-> Double
typeHolderExampleNumberItem

-> Float
typeHolderExampleFloatItem

-> Int
typeHolderExampleIntegerItem

-> Bool
typeHolderExampleBoolItem

-> [Int]
typeHolderExampleArrayItem

-> TypeHolderExample

Construct a value of type TypeHolderExample (by applying it's required fields, if any)

User

data User

User

Constructors

User

Fields

userId :: !(Maybe Integer)

"id"

userUsername :: !(Maybe Text)

"username"

userFirstName :: !(Maybe Text)

"firstName"

userLastName :: !(Maybe Text)

"lastName"

userEmail :: !(Maybe Text)

"email"

userPassword :: !(Maybe Text)

"password"

userPhone :: !(Maybe Text)

"phone"

userUserStatus :: !(Maybe Int)

"userStatus" - User Status

mkUser :: User

Construct a value of type User (by applying it's required fields, if any)

XmlItem

data XmlItem

XmlItem

Constructors

XmlItem

Fields

xmlItemAttributeString :: !(Maybe Text)

"attribute_string"

xmlItemAttributeNumber :: !(Maybe Double)

"attribute_number"

xmlItemAttributeInteger :: !(Maybe Int)

"attribute_integer"

xmlItemAttributeBoolean :: !(Maybe Bool)

"attribute_boolean"

xmlItemWrappedArray :: !(Maybe [Int])

"wrapped_array"

xmlItemNameString :: !(Maybe Text)

"name_string"

xmlItemNameNumber :: !(Maybe Double)

"name_number"

xmlItemNameInteger :: !(Maybe Int)

"name_integer"

xmlItemNameBoolean :: !(Maybe Bool)

"name_boolean"

xmlItemNameArray :: !(Maybe [Int])

"name_array"

xmlItemNameWrappedArray :: !(Maybe [Int])

"name_wrapped_array"

xmlItemPrefixString :: !(Maybe Text)

"prefix_string"

xmlItemPrefixNumber :: !(Maybe Double)

"prefix_number"

xmlItemPrefixInteger :: !(Maybe Int)

"prefix_integer"

xmlItemPrefixBoolean :: !(Maybe Bool)

"prefix_boolean"

xmlItemPrefixArray :: !(Maybe [Int])

"prefix_array"

xmlItemPrefixWrappedArray :: !(Maybe [Int])

"prefix_wrapped_array"

xmlItemNamespaceString :: !(Maybe Text)

"namespace_string"

xmlItemNamespaceNumber :: !(Maybe Double)

"namespace_number"

xmlItemNamespaceInteger :: !(Maybe Int)

"namespace_integer"

xmlItemNamespaceBoolean :: !(Maybe Bool)

"namespace_boolean"

xmlItemNamespaceArray :: !(Maybe [Int])

"namespace_array"

xmlItemNamespaceWrappedArray :: !(Maybe [Int])

"namespace_wrapped_array"

xmlItemPrefixNsString :: !(Maybe Text)

"prefix_ns_string"

xmlItemPrefixNsNumber :: !(Maybe Double)

"prefix_ns_number"

xmlItemPrefixNsInteger :: !(Maybe Int)

"prefix_ns_integer"

xmlItemPrefixNsBoolean :: !(Maybe Bool)

"prefix_ns_boolean"

xmlItemPrefixNsArray :: !(Maybe [Int])

"prefix_ns_array"

xmlItemPrefixNsWrappedArray :: !(Maybe [Int])

"prefix_ns_wrapped_array"

mkXmlItem :: XmlItem

Construct a value of type XmlItem (by applying it's required fields, if any)

Enums

E'ArrayEnum

data E'ArrayEnum

Enum of Text

Constructors

E'ArrayEnum'Fish
"fish"

E'ArrayEnum'Crab
"crab"

fromE'ArrayEnum :: E'ArrayEnum -> Text

unwrap E'ArrayEnum enum

toE'ArrayEnum :: Text -> Either String E'ArrayEnum

parse E'ArrayEnum enum

E'EnumFormString

data E'EnumFormString

Enum of Text . Form parameter enum test (string)

Constructors

E'EnumFormString'_abc
"_abc"

E'EnumFormString'_efg
"-efg"

E'EnumFormString'_xyz
"(xyz)"

fromE'EnumFormString :: E'EnumFormString -> Text

unwrap E'EnumFormString enum

toE'EnumFormString :: Text -> Either String E'EnumFormString

parse E'EnumFormString enum

E'EnumFormStringArray

data E'EnumFormStringArray

Enum of Text

Constructors

E'EnumFormStringArray'GreaterThan
">"

E'EnumFormStringArray'Dollar
"$"

fromE'EnumFormStringArray :: E'EnumFormStringArray -> Text

unwrap E'EnumFormStringArray enum

toE'EnumFormStringArray :: Text -> Either String E'EnumFormStringArray

parse E'EnumFormStringArray enum

E'EnumInteger

data E'EnumInteger

Enum of Int

Constructors

E'EnumInteger'Num1
1

E'EnumInteger'NumMinus_1
-1

fromE'EnumInteger :: E'EnumInteger -> Int

unwrap E'EnumInteger enum

toE'EnumInteger :: Int -> Either String E'EnumInteger

parse E'EnumInteger enum

E'EnumNumber

data E'EnumNumber

Enum of Double

Constructors

E'EnumNumber'Num1_Dot_1
1.1

E'EnumNumber'NumMinus_1_Dot_2
-1.2

fromE'EnumNumber :: E'EnumNumber -> Double

unwrap E'EnumNumber enum

toE'EnumNumber :: Double -> Either String E'EnumNumber

parse E'EnumNumber enum

E'EnumQueryInteger

data E'EnumQueryInteger

Enum of Int

Constructors

E'EnumQueryInteger'Num1
1

E'EnumQueryInteger'NumMinus_2
-2

fromE'EnumQueryInteger :: E'EnumQueryInteger -> Int

unwrap E'EnumQueryInteger enum

toE'EnumQueryInteger :: Int -> Either String E'EnumQueryInteger

parse E'EnumQueryInteger enum

E'EnumString

data E'EnumString

Enum of Text

Constructors

E'EnumString'UPPER
UPPER

E'EnumString'Lower
"lower"

E'EnumString'Empty
""

fromE'EnumString :: E'EnumString -> Text

unwrap E'EnumString enum

toE'EnumString :: Text -> Either String E'EnumString

parse E'EnumString enum

E'Inner

data E'Inner

Enum of Text

Constructors

E'Inner'UPPER
UPPER

E'Inner'Lower
"lower"

fromE'Inner :: E'Inner -> Text

unwrap E'Inner enum

toE'Inner :: Text -> Either String E'Inner

parse E'Inner enum

E'JustSymbol

data E'JustSymbol

Enum of Text

Constructors

E'JustSymbol'Greater_Than_Or_Equal_To
">="

E'JustSymbol'Dollar
"$"

fromE'JustSymbol :: E'JustSymbol -> Text

unwrap E'JustSymbol enum

toE'JustSymbol :: Text -> Either String E'JustSymbol

parse E'JustSymbol enum

E'Kind

data E'Kind

Enum of Text

Constructors

E'Kind'Lions
"lions"

E'Kind'Tigers
"tigers"

E'Kind'Leopards
"leopards"

E'Kind'Jaguars
"jaguars"

fromE'Kind :: E'Kind -> Text

unwrap E'Kind enum

toE'Kind :: Text -> Either String E'Kind

parse E'Kind enum

E'Status

data E'Status

Enum of Text . Order Status

Constructors

E'Status'Placed
"placed"

E'Status'Approved
"approved"

E'Status'Delivered
"delivered"

fromE'Status :: E'Status -> Text

unwrap E'Status enum

toE'Status :: Text -> Either String E'Status

parse E'Status enum

E'Status2

data E'Status2

Enum of Text . pet status in the store

Constructors

E'Status2'Available
"available"

E'Status2'Pending
"pending"

E'Status2'Sold
"sold"

fromE'Status2 :: E'Status2 -> Text

unwrap E'Status2 enum

toE'Status2 :: Text -> Either String E'Status2

parse E'Status2 enum

EnumClass

data EnumClass

Enum of Text

Constructors

EnumClass'_abc
"_abc"

EnumClass'_efg
"-efg"

EnumClass'_xyz
"(xyz)"

fromEnumClass :: EnumClass -> Text

unwrap EnumClass enum

toEnumClass :: Text -> Either String EnumClass

parse EnumClass enum

OuterEnum

data OuterEnum

Enum of Text

Constructors

OuterEnum'Placed
"placed"

OuterEnum'Approved
"approved"

OuterEnum'Delivered
"delivered"

fromOuterEnum :: OuterEnum -> Text

unwrap OuterEnum enum

toOuterEnum :: Text -> Either String OuterEnum

parse OuterEnum enum

Auth Methods

AuthApiKeyApiKey

data AuthApiKeyApiKey

Constructors

AuthApiKeyApiKey Text
secret

AuthApiKeyApiKeyQuery

data AuthApiKeyApiKeyQuery

Constructors

AuthApiKeyApiKeyQuery Text
secret

AuthBasicHttpBasicTest

data AuthBasicHttpBasicTest

Constructors

AuthBasicHttpBasicTest ByteString ByteString
username password

AuthOAuthPetstoreAuth

data AuthOAuthPetstoreAuth

Constructors

AuthOAuthPetstoreAuth Text
secret