OpenAPIPetstore.API.Pet

Description

Operations

Pet

addPet

addPet

Arguments

:: (Consumes AddPet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> OpenAPIPetstoreRequest AddPet contentType NoContent MimeNoContent

POST /pet

Add a new pet to the store

AuthMethod: AuthOAuthPetstoreAuth

data AddPet

deletePet

deletePet

Arguments

:: PetId
"petId" - Pet id to delete

-> OpenAPIPetstoreRequest DeletePet MimeNoContent NoContent MimeNoContent

DELETE /pet/{petId}

Deletes a pet

AuthMethod: AuthOAuthPetstoreAuth

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: Accept accept
request accept (MimeType)

-> Status
"status" - Status values that need to be considered for filter

-> OpenAPIPetstoreRequest FindPetsByStatus MimeNoContent [Pet] accept

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: Accept accept
request accept (MimeType)

-> Tags
"tags" - Tags to filter by

-> OpenAPIPetstoreRequest FindPetsByTags MimeNoContent [Pet] accept

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByTags

getPetById

getPetById

Arguments

:: Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet to return

-> OpenAPIPetstoreRequest GetPetById MimeNoContent Pet accept

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: AuthApiKeyApiKey

data GetPetById

updatePet

updatePet

Arguments

:: (Consumes UpdatePet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> OpenAPIPetstoreRequest UpdatePet contentType NoContent MimeNoContent

PUT /pet

Update an existing pet

AuthMethod: AuthOAuthPetstoreAuth

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Consumes UpdatePetWithForm MimeFormUrlEncoded

=> PetId
"petId" - ID of pet that needs to be updated

-> OpenAPIPetstoreRequest UpdatePetWithForm MimeFormUrlEncoded NoContent MimeNoContent

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: AuthOAuthPetstoreAuth

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Consumes UploadFile MimeMultipartFormData

=> PetId
"petId" - ID of pet to update

-> OpenAPIPetstoreRequest UploadFile MimeMultipartFormData ApiResponse MimeJSON

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: AuthOAuthPetstoreAuth

data UploadFile

uploadFileWithRequiredFile

uploadFileWithRequiredFile

Arguments

:: Consumes UploadFileWithRequiredFile MimeMultipartFormData

=> RequiredFile
"requiredFile" - file to upload

-> PetId
"petId" - ID of pet to update

-> OpenAPIPetstoreRequest UploadFileWithRequiredFile MimeMultipartFormData ApiResponse MimeJSON

POST /fake/{petId}/uploadImageWithRequiredFile

uploads an image (required)

AuthMethod: AuthOAuthPetstoreAuth

data UploadFileWithRequiredFile